backend/requirements.txt
absl-py==0.7.0
astor==0.7.1
Click==7.0
cycler==0.10.0
Django==2.1.7
django-cors-headers==2.5.0
django-filter==2.1.0
djangorestframework==3.9.2
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
html5lib==1.0.1
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==2.0.0
mod-wsgi==4.6.5
numpy==1.16.2
pandas==0.24.1
pbr==5.1.3
Pillow==5.4.1
protobuf==3.7.0
psycopg2-binary==2.7.7
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
virtualenv==16.4.3
webencodings==0.5.1
Werkzeug==0.14.1
